Moved precacheScreenshots thunk out of protocol/graphics and into actions/timeline
